CHROMIUM: mfd: cros_ec: add debugfs, console log file
If the EC supports the new CONSOLE_READ command type, then we
place a console_log file in debugfs for that EC device which allows
us to grab EC logs. The kernel will poll every 30 seconds for the
log and keep its own buffer, but userspace should grab this and
write it out to some logs which actually get rotated.

[drinkcat: Cherry-picked from 3.14 tree,
commit 62d462b30955 ("mfd/cros_ec: add debugfs, console log file")

Locations and names of structs changed, some bug fixes, moved
away from seq_file so that we can implement polling more
properly.]

BUG=chromium:492721
BUG=chromium:527904
TEST=on oak, cat /sys/kernel/debug/cros_ec/console_log and see
  that it has data, cat again and see that it blocks until the
  kernel has polled the EC and data is available

Change-Id: Icac784ed3c63b6963d23cad9958aa86507d79d0a
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299804
Reviewed-by: Guenter Roeck <groeck@google.com>