UPSTREAM: PM / OPP: Break _opp_add_dynamic() into smaller functions
Later commits would add support for new OPP bindings and this would be
required then. So, lets do it in a separate patch to make it easily
reviewable.

Another change worth noticing is INIT_LIST_HEAD(&opp->node). We weren't
doing it earlier as we never tried to delete a list node before it is
added to list. But this wouldn't be the case anymore. We might try to
delete a node (just to reuse the same code paths), without it being
getting added to the list.

Reviewed-by: Bartlomiej Zolnierkiewicz <b.zolnierkie@samsung.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 23dacf6d2e993551ef3ae0629baf6f473930513c)
Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@mediatek.com>

BUG=chrome-os-partner:50411
TEST=build and boot on oak-rev5

Change-Id: Ib194132c08b7758c00e3a02f902a9894b7c720fe
Reviewed-on: https://chromium-review.googlesource.com/330321
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>