BACKPORT: PM / sleep: Go direct_complete if driver has no callbacks
If a suitable prepare callback cannot be found for a given device and
its driver has no PM callbacks at all, assume that it can go direct to
complete when the system goes to sleep.

The reason for this is that there's lots of devices in a system that do
no PM at all and there's no reason for them to prevent their ancestors
to do direct_complete if they can support it.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit aa8e54b559479d0cb7eb632ba443b8cacd20cd4b)

Conflicts:
	include/linux/pm.h

BUG=chrome-os-partner:46491
TEST=suspend/resume on glados/chell

Change-Id: Idc5774fa3f237a214af3aac70078a4f6839b3afd
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326882
Reviewed-by: Benson Leung <bleung@chromium.org>