drm/panel: lpm102a188a: Avoid backlight updates when panel is disabled
Call to turn off panel (panel_jdi_disable()/panel_jdi_unprepare()) are
not protected by lock. If there is a backlight update
(backlight_jdi_update_status()) this update may go unnoticed and
ignored. If panel is turned back on the backlight may not be turned
back to the correct value because of this. Addressing this by
disabling backlight updates once the panel off sequence begins.

BUG: 29345241
Change-Id: Ib5e71950b0fb22bf4659b6063b55e4903cef1ce5
Signed-off-by: Adrian Salido <salidoa@google.com>