BACKPORT: net: add recursion limit to GRO
Currently, GRO can do unlimited recursion through the gro_receive
handlers.  This was fixed for tunneling protocols by limiting tunnel GRO
to one level with encap_mark, but both VLAN and TEB still have this
problem.  Thus, the kernel is vulnerable to a stack overflow, if we
receive a packet composed entirely of VLAN headers.

This patch adds a recursion counter to the GRO layer to prevent stack
overflow.  When a gro_receive function hits the recursion limit, GRO is
aborted for this skb and it is processed normally.  This recursion
counter is put in the GRO CB, but could be turned into a percpu counter
if we run out of space in the CB.

Thanks to Vladimr Bene <vbenes@redhat.com> for the initial bug report.

Fixes: CVE-2016-7039
Fixes: 9b174d88c257 ("net: Add Transparent Ethernet Bridging GRO support.")
Fixes: 66e5133f19e9 ("vlan: Add GRO support for non hardware accelerated vlan")
Signed-off-by: Sabrina Dubroca <sd@queasysnail.net>
Reviewed-by: Jiri Benc <jbenc@redhat.com>
Acked-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Acked-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

(cherry picked from commit fcd91dd449867c6bfe56a81cabba76b829fd05cd)

Notes on backport:
------------------------------------------------------------------------
The upstream patch doesn't apply cleanly on top of ChromeOS 3.18 kernel.

- The call_gro_receive change doesn't apply to the following files
  as no call to .gro_receive exists in the code yet.

     drivers/net/geneve.c
     net/8021q/vlan.c
     net/ethernet/eth.c

- call_gro_receive is used instead of call_gro_receive_sk in
  net/ipv4/udp_offload.c as the .gro_receive callback doesn't take
  `struct sk` argument.
------------------------------------------------------------------------

BUG=b:32515955, chromium:672642
TEST=Trybot runs on paladin and release builders.

Change-Id: Ica125ce3a3d30633c3b74c37676b966bdbc4b47b
Signed-off-by: Ben Chan <benchan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/419721
Reviewed-by: Mattias Nissler <mnissler@chromium.org>