Revert "UPSTREAM: ecryptfs: forbid opening files without mmap handler"
This reverts commit 564647162b0f3adc885c3a63e8dfbe77d75d791d.

It fixed a local root exploit but also introduced a dependency on
the lower file system implementing an mmap operation just to open a file,
which is a bit of a heavy hammer.  The right fix is to have mmap depend
on the existence of the mmap handler instead.

Conflicts:
	fs/ecryptfs/kthread.c

Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>