CHROMIUM: md: verity: Properly handle verity version
In its current state, DM_VERITY_NUM_POSITIONAL_ARGS was used to check
whether to use the original ChromeOS verity or the present verity. This is
utterly broken since this does not exclude additional mount options
appended to the arguments when determining the number of arguments.

Instead of using DM_VERITY_NUM_POSITIONAL_ARGS, retrieve and validate the
version argument, and use it to determine which variant of verity to use.

BUG=none
TEST= Kernel compiles and boots smaug \
      >> dmesg | grep dm-0
[    2.403547] EXT4-fs (dm-0): mounted filesystem with ordered data mode. Opts: barrier=1

Tested-by: Aun-Ali Zaidi <admin@kodeit.net>
Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>